pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "memory-picker"
version = "0.1.0"
description = "Deterministic local photo intake and filtering pipeline for trip albums."
readme = "instructions.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.0",
    "pillow>=11.1.0",
    "pillow-heif>=0.21.0",
]

[project.scripts]
memory-picker = "memory_picker.cli:main"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/memory_picker"]
